Split stake api handlers into stake.go
